# backend/docs/swagger.yaml
definitions:
  models.Affiliate:
    properties:
      code:
        type: string
      commissionRate:
        type: number
      earned:
        type: number
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      pixKey:
        type: string
      referrals:
        type: integer
      status:
        type: string
    type: object
  models.Client:
    properties:
      casa_alugada:
        type: boolean
      casa_propria:
        type: boolean
      city:
        type: string
      comprovante_residencia:
        type: boolean
      cpf:
        type: string
      data_nascimento:
        type: string
      email:
        type: string
      empresa:
        type: string
      endereco:
        type: string
      estuda:
        type: boolean
      filhos:
        type: integer
      id: {}
      indicado:
        type: string
      instagram:
        type: string
      modelo_trabalho:
        type: string
      name:
        type: string
      nome_sujo:
        type: boolean
      outro_emprestimo:
        type: integer
      phone:
        type: string
      profissao:
        type: string
      salario:
        type: integer
      status:
        type: string
      tempo_empresa:
        type: string
      vale_alimentacao:
        type: boolean
      vale_refeicao:
        type: boolean
      valor_emprestimo:
        type: integer
      valor_esperado:
        type: integer
      valor_outro_emprestimo:
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: mensagem de erro
        type: string
    type: object
  models.Loan:
    properties:
      amount:
        type: number
      checklistAtApproval:
        items:
          type: string
        type: array
      client:
        type: string
      fineRate:
        type: number
      history:
        items:
          $ref: '#/definitions/models.PaymentRecord'
        type: array
      id:
        type: string
      installmentValue:
        type: number
      installments:
        type: integer
      interestRate:
        type: number
      justification:
        type: string
      moraInterestRate:
        type: number
      nextDue:
        type: string
      startDate:
        type: string
      status:
        type: string
      totalPaidCapital:
        type: number
      totalPaidInterest:
        type: number
    type: object
  models.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: admin
        type: string
    type: object
  models.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
  models.PaymentRecord:
    properties:
      amount:
        type: number
      date:
        type: string
      note:
        type: string
      type:
        type: string
    type: object
  models.SuccessMessage:
    properties:
      message:
        example: operacao realizada com sucesso
        type: string
    type: object
  models.User:
    properties:
      id: {}
      nome:
        type: string
      password:
        type: string
      telefone:
        type: string
      username:
        type: string
    type: object
  models.UsuariosResponse:
    properties:
      usuarios:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
info:
  contact: {}
paths:
  /affiliates:
    get:
      description: Retorna todos os afiliados cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Affiliate'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar afiliados
      tags:
      - Affiliates
    post:
      consumes:
      - application/json
      description: Cadastra um novo afiliado
      parameters:
      - description: Dados do afiliado
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Affiliate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Affiliate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar afiliado
      tags:
      - Affiliates
  /affiliates/{id}:
    delete:
      description: Remove um afiliado pelo ID
      parameters:
      - description: ID do afiliado
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletar afiliado
      tags:
      - Affiliates
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um afiliado pelo ID
      parameters:
      - description: ID do afiliado
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Affiliate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Affiliate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar afiliado
      tags:
      - Affiliates
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica o usuario com username e password e retorna um token
        JWT
      parameters:
      - description: Credenciais
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /customers:
    get:
      description: Retorna todos os clientes cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Client'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar clientes
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Cadastra um novo cliente
      parameters:
      - description: Dados do cliente
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar cliente
      tags:
      - Customers
  /customers/{id}:
    delete:
      description: Remove um cliente pelo ID
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletar cliente
      tags:
      - Customers
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um cliente pelo ID
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar cliente
      tags:
      - Customers
  /loans:
    get:
      description: Retorna todos os emprestimos cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Loan'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar emprestimos
      tags:
      - Loans
    post:
      consumes:
      - application/json
      description: Cadastra um novo emprestimo
      parameters:
      - description: Dados do emprestimo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Loan'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Loan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar emprestimo
      tags:
      - Loans
  /loans/{id}:
    delete:
      description: Remove um emprestimo pelo ID
      parameters:
      - description: ID do emprestimo
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletar emprestimo
      tags:
      - Loans
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um emprestimo pelo ID
      parameters:
      - description: ID do emprestimo
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Loan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Loan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar emprestimo
      tags:
      - Loans
  /users:
    get:
      description: Retorna a lista de todos os usuarios cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsuariosResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar usuarios
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Cadastra um novo usuario no sistema
      parameters:
      - description: Dados do usuario
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar usuario
      tags:
      - Users
  /users/{id}:
    delete:
      description: Remove um usuario pelo ID
      parameters:
      - description: ID do usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletar usuario
      tags:
      - Users
    get:
      description: Retorna um usuario pelo seu ID
      parameters:
      - description: ID do usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar usuario por ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um usuario pelo ID
      parameters:
      - description: ID do usuario
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar usuario
      tags:
      - Users
swagger: "2.0"